+++ src/FM3_MicRo/data_plotting_scripts/npz_plotter.py
@@ -87,7 +87,7 @@ class NPZPlotter:
         plt.title("Evaluation Results")
         plt.legend()
         plt.grid()
-        plt.ylim(bottom=min(self.results,0))
+        plt.ylim(bottom=min(np.min(self.results),0))
 
 
 if __name__ == "__main__":